requirements.txt
accelerate
bitsandbytes  
fastapi
uvicorn
pymongo
python-dotenv
transformers
torch